+++ marsworks/origin/serializer.py
@@ -9,17 +9,31 @@ __all__ = ("Serializer",)
 
 
 class Serializer:
+    """
+    A class representing a Serializer, used for serializing response into
+    other objects.
 
-    __slots__ = ("_response",)
+    Attributes:
+        response (httpx.Response): The response API returned.
+
+    Warning:
+        This object is not for public use unless `await Client.get_raw_response()`
+        is being used.
+    """
+
+    __slots__ = ("response",)
 
     def __init__(self, response: httpx.Response) -> None:
-        self._response = response
+        self.response = response
 
     async def manifest_content(self) -> Manifest:
         """
-        Serializes into Manifest.
+        Serializes into [Manifest](./manifest.md).
+
+        Returns:
+            A [Manifest](./manifest.md) object containing mission's info.
         """
-        data = (self._response.json())["rover"]
+        data = (self.response.json())["rover"]
         if data:
             return Manifest(data)
         else:
@@ -27,9 +41,12 @@ class Serializer:
 
     async def photo_content(self) -> list:
         """
-        Serializes into Photo.
+        Serializes into a list of [Photo](./photo.md).
+
+        Returns:
+            A list of [Photo](./photo.md) objects with url and info.
         """
-        data = self._response.json()
+        data = self.response.json()
         for key in data:
             if key in ("photos", "latest_photos"):
                 data = data[key]